+++ src/errors/InvalidDirectiveError.ts
@@ -1,0 +1,6 @@
+export class InvalidDirectiveError extends Error {
+  constructor(msg: string) {
+    super(msg);
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}