+++ src/errors/index.ts
@@ -5,6 +5,7 @@ export * from "./GeneratingSchemaError";
 export * from "./ConflictingDefaultValuesError";
 export * from "./ConflictingDefaultWithNullableError";
 export * from "./InterfaceResolveTypeError";
+export * from "./InvalidDirectiveError";
 export * from "./MissingSubscriptionTopicsError";
 export * from "./NoExplicitTypeError";
 export * from "./ReflectMetadataMissingError";