+++ src/metadata/definitions/directive-metadata.ts
@@ -1,0 +1,15 @@
+export interface DirectiveMetadata {
+  nameOrDefinition: string;
+  args: Record<string, any>;
+}
+
+export interface DirectiveClassMetadata {
+  target: Function;
+  directive: DirectiveMetadata;
+}
+
+export interface DirectiveFieldMetadata {
+  target: Function;
+  fieldName: string;
+  directive: DirectiveMetadata;
+}