+++ src/schema/schema-generator.ts
@@ -15,6 +15,7 @@ import {
   GraphQLEnumValueConfigMap,
   GraphQLUnionType,
   GraphQLTypeResolver,
+  GraphQLDirective,
 } from "graphql";
 import { withFilter, ResolverFn } from "graphql-subscriptions";
 
@@ -43,6 +44,12 @@ import {
 import { ResolverFilterData, ResolverTopicData, TypeResolver } from "../interfaces";
 import { getFieldMetadataFromInputType, getFieldMetadataFromObjectType } from "./utils";
 import { ensureInstalledCorrectGraphQLPackage } from "../utils/graphql-version";
+import {
+  getFieldDefinitionNode,
+  getInputObjectTypeDefinitionNode,
+  getInputValueDefinitionNode,
+  getObjectTypeDefinitionNode,
+} from "./definition-node";
 
 interface AbstractInfo {
   isAbstract: boolean;
@@ -81,6 +88,11 @@ export interface SchemaGeneratorOptions extends BuildContextOptions {
    * Disable checking on build the correctness of a schema
    */
   skipCheck?: boolean;
+
+  /**
+   * Array of graphql directives
+   */
+  directives?: GraphQLDirective[];
 }
 
 export abstract class SchemaGenerator {
@@ -113,6 +125,7 @@ export abstract class SchemaGenerator {
       mutation: this.buildRootMutationType(options.resolvers),
       subscription: this.buildRootSubscriptionType(options.resolvers),
       types: this.buildOtherTypes(orphanedTypes),
+      directives: options.directives,
     });
 
     BuildContext.reset();
@@ -268,6 +281,7 @@ export abstract class SchemaGenerator {
         type: new GraphQLObjectType({
           name: objectType.name,
           description: objectType.description,
+          astNode: getObjectTypeDefinitionNode(objectType.name, objectType.directives),
           interfaces: () => {
             let interfaces = interfaceClasses.map<GraphQLInterfaceType>(
               interfaceClass =>
@@ -293,14 +307,20 @@ export abstract class SchemaGenerator {
                     (resolver.resolverClassMetadata === undefined ||
                       resolver.resolverClassMetadata.isAbstract === false),
                 );
+                const type = this.getGraphQLOutputType(
+                  field.name,
+                  field.getType(),
+                  field.typeOptions,
+                );
                 fieldsMap[field.schemaName] = {
-                  type: this.getGraphQLOutputType(field.name, field.getType(), field.typeOptions),
+                  type,
                   args: this.generateHandlerArgs(field.params!),
                   resolve: fieldResolverMetadata
                     ? createAdvancedFieldResolver(fieldResolverMetadata)
                     : createSimpleFieldResolver(field),
                   description: field.description,
                   deprecationReason: field.deprecationReason,
+                  astNode: getFieldDefinitionNode(field.name, type, field.directives),
                   extensions: {
                     complexity: field.complexity,
                   },
@@ -361,10 +381,16 @@ export abstract class SchemaGenerator {
                   inputType.name,
                 );
 
+                const type = this.getGraphQLInputType(
+                  field.name,
+                  field.getType(),
+                  field.typeOptions,
+                );
                 fieldsMap[field.schemaName] = {
                   description: field.description,
-                  type: this.getGraphQLInputType(field.name, field.getType(), field.typeOptions),
+                  type,
                   defaultValue: field.typeOptions.defaultValue,
+                  astNode: getInputValueDefinitionNode(field.name, type, field.directives),
                 };
                 return fieldsMap;
               },
@@ -380,6 +406,7 @@ export abstract class SchemaGenerator {
             }
             return fields;
           },
+          astNode: getInputObjectTypeDefinitionNode(inputType.name, inputType.directives),
         }),
       };
     });
@@ -449,16 +476,18 @@ export abstract class SchemaGenerator {
       if (handler.resolverClassMetadata && handler.resolverClassMetadata.isAbstract) {
         return fields;
       }
+      const type = this.getGraphQLOutputType(
+        handler.methodName,
+        handler.getReturnType(),
+        handler.returnTypeOptions,
+      );
       fields[handler.schemaName] = {
-        type: this.getGraphQLOutputType(
-          handler.methodName,
-          handler.getReturnType(),
-          handler.returnTypeOptions,
-        ),
+        type,
         args: this.generateHandlerArgs(handler.params!),
         resolve: createHandlerResolver(handler),
         description: handler.description,
         deprecationReason: handler.deprecationReason,
+        astNode: getFieldDefinitionNode(handler.schemaName, type, handler.directives),
         extensions: {
           complexity: handler.complexity,
         },